+++ src/types.rs
@@ -1,247 +1,4 @@
 use chrono::{DateTime, TimeZone, Utc};
-use hyper;
-use std::error::Error;
-use std::fmt;
-use std::io;
-use std::str::FromStr;
-
-#[derive(Deserialize, Debug)]
-pub struct JsonError {
-    pub error: String,
-    pub error_description: Option<String>,
-    pub error_uri: Option<String>,
-}
-
-/// A helper type to deserialize either a JsonError or another piece of data.
-#[derive(Deserialize, Debug)]
-#[serde(untagged)]
-pub enum JsonErrorOr<T> {
-    Err(JsonError),
-    Data(T),
-}
-
-/// All possible outcomes of the refresh flow
-#[derive(Debug)]
-pub enum RefreshResult {
-    /// Indicates connection failure
-    Error(hyper::Error),
-    /// The server did not answer with a new token, providing the server message
-    RefreshError(String, Option<String>),
-    /// The refresh operation finished successfully, providing a new `Token`
-    Success(Token),
-}
-
-/// Encapsulates all possible results of a `poll_token(...)` operation in the Device flow.
-#[derive(Debug)]
-pub enum PollError {
-    /// Connection failure - retry if you think it's worth it
-    HttpError(hyper::Error),
-    /// Indicates we are expired, including the expiration date
-    Expired(DateTime<Utc>),
-    /// Indicates that the user declined access. String is server response
-    AccessDenied,
-    /// Indicates that too many attempts failed.
-    TimedOut,
-    /// Other type of error.
-    Other(String),
-}
-
-/// Encapsulates all possible results of the `token(...)` operation
-#[derive(Debug)]
-pub enum RequestError {
-    /// Indicates connection failure
-    ClientError(hyper::Error),
-    /// The OAuth client was not found
-    InvalidClient,
-    /// Some requested scopes were invalid. String contains the scopes as part of
-    /// the server error message
-    InvalidScope(String),
-    /// A 'catch-all' variant containing the server error and description
-    /// First string is the error code, the second may be a more detailed description
-    NegativeServerResponse(String, Option<String>),
-    /// A malformed server response.
-    BadServerResponse(String),
-    /// Error while decoding a JSON response.
-    JSONError(serde_json::Error),
-    /// Error within user input.
-    UserError(String),
-    /// A lower level IO error.
-    LowLevelError(io::Error),
-    /// A poll error occurred in the DeviceFlow.
-    Poll(PollError),
-    /// An error occurred while refreshing tokens.
-    Refresh(RefreshResult),
-    /// Error in token cache layer
-    Cache(Box<dyn Error + Send + Sync>),
-}
-
-impl From<hyper::Error> for RequestError {
-    fn from(error: hyper::Error) -> RequestError {
-        RequestError::ClientError(error)
-    }
-}
-
-impl From<JsonError> for RequestError {
-    fn from(value: JsonError) -> RequestError {
-        match &*value.error {
-            "invalid_client" => RequestError::InvalidClient,
-            "invalid_scope" => RequestError::InvalidScope(
-                value
-                    .error_description
-                    .unwrap_or_else(|| "no description provided".to_string()),
-            ),
-            _ => RequestError::NegativeServerResponse(value.error, value.error_description),
-        }
-    }
-}
-
-impl From<serde_json::Error> for RequestError {
-    fn from(value: serde_json::Error) -> RequestError {
-        RequestError::JSONError(value)
-    }
-}
-
-impl fmt::Display for RequestError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        match *self {
-            RequestError::ClientError(ref err) => err.fmt(f),
-            RequestError::InvalidClient => "Invalid Client".fmt(f),
-            RequestError::InvalidScope(ref scope) => writeln!(f, "Invalid Scope: '{}'", scope),
-            RequestError::NegativeServerResponse(ref error, ref desc) => {
-                error.fmt(f)?;
-                if let Some(ref desc) = *desc {
-                    write!(f, ": {}", desc)?;
-                }
-                "\n".fmt(f)
-            }
-            RequestError::BadServerResponse(ref s) => s.fmt(f),
-            RequestError::JSONError(ref e) => format!(
-                "JSON Error; this might be a bug with unexpected server responses! {}",
-                e
-            )
-            .fmt(f),
-            RequestError::UserError(ref s) => s.fmt(f),
-            RequestError::LowLevelError(ref e) => e.fmt(f),
-            RequestError::Poll(ref pe) => pe.fmt(f),
-            RequestError::Refresh(ref rr) => format!("{:?}", rr).fmt(f),
-            RequestError::Cache(ref e) => e.fmt(f),
-        }
-    }
-}
-
-impl Error for RequestError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match *self {
-            RequestError::ClientError(ref err) => Some(err),
-            RequestError::LowLevelError(ref err) => Some(err),
-            RequestError::JSONError(ref err) => Some(err),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct StringError {
-    error: String,
-}
-
-impl fmt::Display for StringError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        self.description().fmt(f)
-    }
-}
-
-impl StringError {
-    pub fn new<S: AsRef<str>>(error: S, desc: Option<S>) -> StringError {
-        let mut error = error.as_ref().to_string();
-        if let Some(d) = desc {
-            error.push_str(": ");
-            error.push_str(d.as_ref());
-        }
-
-        StringError { error }
-    }
-}
-
-impl<'a> From<&'a dyn Error> for StringError {
-    fn from(err: &'a dyn Error) -> StringError {
-        StringError::new(err.description().to_string(), None)
-    }
-}
-
-impl From<String> for StringError {
-    fn from(value: String) -> StringError {
-        StringError::new(value, None)
-    }
-}
-
-impl Error for StringError {
-    fn description(&self) -> &str {
-        &self.error
-    }
-}
-
-/// Represents all implemented token types
-#[derive(Clone, PartialEq, Debug)]
-pub enum TokenType {
-    /// Means that whoever bears the access token will be granted access
-    Bearer,
-}
-
-impl AsRef<str> for TokenType {
-    fn as_ref(&self) -> &'static str {
-        match *self {
-            TokenType::Bearer => "Bearer",
-        }
-    }
-}
-
-impl FromStr for TokenType {
-    type Err = ();
-    fn from_str(s: &str) -> Result<TokenType, ()> {
-        match s {
-            "Bearer" => Ok(TokenType::Bearer),
-            _ => Err(()),
-        }
-    }
-}
-
-/// A scheme for use in `hyper::header::Authorization`
-#[derive(Clone, PartialEq, Debug)]
-pub struct Scheme {
-    /// The type of our access token
-    pub token_type: TokenType,
-    /// The token returned by one of the Authorization Flows
-    pub access_token: String,
-}
-
-impl std::convert::Into<hyper::header::HeaderValue> for Scheme {
-    fn into(self) -> hyper::header::HeaderValue {
-        hyper::header::HeaderValue::from_str(&format!(
-            "{} {}",
-            self.token_type.as_ref(),
-            self.access_token
-        ))
-        .expect("Invalid Scheme header value")
-    }
-}
-
-impl FromStr for Scheme {
-    type Err = &'static str;
-    fn from_str(s: &str) -> Result<Scheme, &'static str> {
-        let parts: Vec<&str> = s.split(' ').collect();
-        if parts.len() != 2 {
-            return Err("Expected two parts: <token_type> <token>");
-        }
-        match <TokenType as FromStr>::from_str(parts[0]) {
-            Ok(t) => Ok(Scheme {
-                token_type: t,
-                access_token: parts[1].to_string(),
-            }),
-            Err(_) => Err("Couldn't parse token type"),
-        }
-    }
-}
 
 /// Represents a token as returned by OAuth2 servers.
 ///
@@ -362,7 +119,6 @@ pub struct ConsoleApplicationSecret {
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use hyper;
 
     pub const SECRET: &'static str =
         "{\"installed\":{\"auth_uri\":\"https://accounts.google.com/o/oauth2/auth\",\
@@ -379,26 +135,5 @@ pub mod tests {
             Ok(s) => assert!(s.installed.is_some() && s.web.is_none()),
             Err(err) => panic!(err),
         }
-    }
-
-    #[test]
-    fn schema() {
-        let s = Scheme {
-            token_type: TokenType::Bearer,
-            access_token: "foo".to_string(),
-        };
-        let mut headers = hyper::HeaderMap::new();
-        headers.insert(hyper::header::AUTHORIZATION, s.into());
-        assert_eq!(
-            format!("{:?}", headers),
-            "{\"authorization\": \"Bearer foo\"}".to_string()
-        );
-    }
-
-    #[test]
-    fn parse_schema() {
-        let auth = Scheme::from_str("Bearer foo").unwrap();
-        assert_eq!(auth.token_type, TokenType::Bearer);
-        assert_eq!(auth.access_token, "foo".to_string());
     }
 }
